requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
youtube-transcript-api>=0.6.1
pydantic>=2.0.0